# Minor changes for performance improvements