This PR features some small changes to some of the operations to increase performance while altering the existing code the least amount possible.

On my machine, this leads to: (`bench-on` is needed because benchmark script not yet merged).

```bash
benchpkg --rev={DEFAULT},dirty --bench-on=2a34e3d8e880469dbfddd9dde18ef34b0d4ebf34 --output-dir=benchmark/results/ 
```

```
|                            | main             | dirty            | main/dirty |
|:---------------------------|:----------------:|:----------------:|:----------:|
| lanczos (N=50)/nterms=2^14 | 0.375 ± 0.02 s   | 0.288 ± 0.013 s  | 1.31       |
| lanczos (N=50)/nterms=2^16 | 1.07 ± 0.024 s   | 0.75 ± 0.048 s   | 1.43       |
| lanczos (N=50)/nterms=2^18 | 3.07 ± 0.019 s   | 2.14 ± 0.033 s   | 1.43       |
| lanczos (N=50)/nterms=2^20 | 6.29 s           | 4.24 ± 0.0083 s  | 1.48       |
| time_to_load               | 0.154 ± 0.0062 s | 0.156 ± 0.0086 s | 0.987      |
```